rename: logo 컴포넌트 공통 컴포넌트로 폴더 변경
logo 컴포넌트 swiper에서만 사용하지 않고 여러군데에서 사용할 것을 감안해 common 컴포넌트 폴더로 변경
