+++ src/components/spot/swiper/SlideItem.tsx
@@ -1,5 +1,6 @@
+import Logo from '@components/common/Logo';
 import styled from '@emotion/styled';
-import Logo from './Logo';
+
 import { Common } from '@styles/globalStyle';
 
 interface Props {
